Finished updating the remaining monte_carlo_collision_core package unit tests using the FRENSIE unit test harness.

+++ packages/monte_carlo/collision/core/src/MonteCarlo_AceLaw61NuclearScatteringDistribution_def.hpp
@@ -52,7 +52,7 @@ void AceLaw61NuclearScatteringDistribution<IncomingParticleType,
                                 OutgoingParticleType& outgoing_particle,
 				const double temperature ) const
 {
-  unsigned outgoing_bin_index, incoming_bin_index;
+  size_t outgoing_bin_index, incoming_bin_index;
   
   double energy_prime;
